+++ src/host/00-namespace.jsx
@@ -5,7 +5,7 @@
  * other panels share this ExtendScript engine.
  */
 var LML = (typeof LML !== "undefined" && LML) ? LML : {};
-LML.version = "0.1.0";
+LML.version = "0.1.1";
 LML.api = LML.api || {};
 LML.util = LML.util || {};